requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.5.1
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.0
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==37.0.2
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.11.0
django-dotenv==1.4.2
django-rest-knox==4.2.0
djangorestframework==3.13.1
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
multidict==6.0.4
openai==0.27.2
Pillow==9.1.0
psycopg2==2.9.3
pycparser==2.21
pytz==2022.1
requests==2.28.2
sqlparse==0.4.2
tqdm==4.65.0
tzdata==2022.1
urllib3==1.26.15
whitenoise==6.0.0
yarl==1.8.2
